+++ course/lesson-26-looping-over-dictionaries/place-units/TestPlaceUnits.gd
@@ -15,9 +15,12 @@ func _define(checks: Array[Check]) -> void:
 func test_all_units_are_displayed():
 	var displayed: Dictionary = grid.get_displayed_units_info()
 	var source: Dictionary = grid.get("units")
+	if displayed.size() != source.size():
+		return tr("The number of units displayed does not match the number of units in the units dictionary. Did you place each unit exactly once?")
+
 	var dicts_match := displayed.has_all(source.keys())
 	if dicts_match:
-		for key in displayed:
+		for key in source:
 			if displayed[key] != source[key]:
 				dicts_match = false
 				break
@@ -29,12 +32,13 @@ func test_all_units_are_displayed():
 
 func test_code_uses_a_for_loop():
 	var run_function := _analyzer.get_function_named("run")
-	
+
 	if not GDExpr.suite(
 		GDExpr.for_loop(
-			GDExpr.any_identifier(),
-			GDExpr.identifier("units")
+			null,
+			null,
+			GDExpr.suite(GDExpr.function_call("place_unit")),
 		)
 	).matches(run_function):
-		return tr("Your code has no for loop. You need to use a for loop to complete this practice, even if there are other solutions!")
+		return tr("Your code does not use a for loop to call place_unit(). Please make sure you place each unit inside the loop.")
 	return ""