+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This document lists changes between releases.
 
+## Unreleased
+
+## Bug fixes
+
+- L25.P1 and L25.P2: Updated practice tests to allow more valid solutions to pass.
+
 ## Learn GDScript From Zero 2.1.0 (2026-08-16)
 
 ## Improvements and changes